+++ test/test_hash_ring.py
@@ -63,31 +63,3 @@ def test_hash_ring_stress_test():
         assert ring.get_num_items() == total_items
 
     assert total_items == 0
-
-
-def ring_fast(num_nodes):
-    r = HashRing()
-    nodes = []
-    for i in xrange(num_nodes):
-        nodes.append(HashRingNode('test-%i' % i, num_replicas=512))
-
-    r.add_nodes(nodes)
-
-
-def ring_slow(num_nodes):
-    r = HashRing()
-    nodes = []
-    for i in xrange(num_nodes):
-        r.add_node('test-%i' % i, num_replicas=512)
-
-    r.add_nodes(nodes)
-
-
-@pytest.mark.parametrize('num_nodes', [5, 10, 25, 50, 100])
-def test_ring_add_nodes_slow(benchmark, num_nodes):
-    benchmark(ring_slow, num_nodes)
-
-
-@pytest.mark.parametrize('num_nodes', [5, 10, 25, 50, 100])
-def test_ring_add_nodes_fast(benchmark, num_nodes):
-    benchmark(ring_fast, num_nodes)
